+++ quantforge/__init__.py
@@ -772,6 +772,7 @@ from .aaa import aaa
 from .prony import prony, prony_reconstruct
 from .matrix_pencil import matrix_pencil
 from .music import music_pseudospectrum, music_frequencies
+from .randomized_svd import randomized_svd
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1476,7 +1477,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.58.0"
+__version__ = "5.59.0"
 
 __all__ = [
     "OptionType",
@@ -2517,6 +2518,7 @@ __all__ = [
     "matrix_pencil",
     "music_pseudospectrum",
     "music_frequencies",
+    "randomized_svd",
     "Dual",
     "dual_derivative",
     "dual_gradient",